refactor ♻️: Refactor AI commit message generation to use new client and prompts
This commit refactors the AI commit message generation process by moving the creation of the AI prompt from `ai_utils.py` to a new file `generation.py`. The `__init__.py` in the `ai` directory is also updated to import the new `generation.py` module. This change enhances code organization and reusability, making it easier to manage and extend future enhancements or changes to the AI generation logic.
