+++ git_acp/ai/__init__.py
@@ -1,5 +1,5 @@
 """AI package for git-acp."""
 
-from git_acp.ai.ai_utils import generate_commit_message
+from git_acp.ai.generation import generate_commit_message
 
 __all__ = ['generate_commit_message']
